Rejigging hero component for more consistency on mobile
Hero component already resized on browser height changes, but the disappearing mobile nav bar is a right royal pain in the ass, and would replot the circle graphics whenever you'd scroll in a direction that'd induce a change to the nav bar.

Anyway. It shouldn't do that any more.

I hope.